Complete ps1 part1
Infinite loop solved and the previous implementation did work to produce the correct output.

However, the online python grade uses python 3.5... which doesn't support dictionary sorting int he way I implemented it.

Modified my dictionary-based implementation to a list-based implementation to resolved this. Both are present in part 1 of this file.

Misc testing code left in place for now, as I might need it for part 2-3.